Small changes
Make vim colors more compatible with dark theme and remove line
numbers. This makes it easier to work with a small terminal window
outside. Plan is to use vim for quick edits rather than as a
source code editor. Use VS Code instead for the LLDB UI and
excellent ssh support.

Add sbin to path in zshrc so that programs installed by brew will
be available.